Wall view now controls game logic for ConnectingWall and Timer child components
